requirements.txt
Django==5.2.8
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==2.0.2
django-filter==24.1
